+++ src/utils/board.js
@@ -1,0 +1,7 @@
+export const getTileCardsArray = (board, place) => {
+  return board[place.y][place.x].cards;
+};
+
+export const getTileCard = (board, place, index) => {
+  return board[place.y][place.x].cards[index];
+};